Messenger/client/requirements.txt
pyyaml
requests
fabric3
tabulate
PyQt5
pycryptodome
sphinx